Neatened details display
Changes to subject.vm: neatened display of Research Themes to just
print the name of the theme that applies rather then lising all themes
and indicating which are "on".
This also involved the creation of a new macro in body.vm.

New file data-download.vm: Gathers into one location the various data
download links that were otherwise scattered through out the various
sections. Only shown to non-logged in users.

Changes to body.vm:
Modified which sections are show to logged in or non-logged in users.
Non-logged in users see data-download.vm; logged-in users see attach.vm
instead.